+++ backend/src/scripts/initPersons.ts
@@ -1,4 +1,4 @@
-import { initPersonDB } from "../services/initService.js";
+import { fetchTMDB } from "./fetchHelper.js";
 
 const apiKey = process.env.API_KEY;
 const delay = 30;
@@ -19,34 +19,15 @@ export interface PersonData {
   profile_path: string;
 }
 
-interface ResponseData extends PersonData {
+interface PersonResponse extends PersonData {
   also_known_as: string[];
 }
 
-const fetchPerson = async (personId: number) => {
-  try {
-    const response = await fetch(`https://api.themoviedb.org/3/person/${personId}?api_key=${apiKey}`);
-
-    if (!response.ok) {
-      throw response.status;
-    }
-
-    const data = (await response.json()) as ResponseData;
-
-    return data;
-  } catch (error) {
-    // console.error(`Error fetching data for person ID ${personId}:`, error.message);
-    if (error === 429) {
-      console.log("429");
-      // await new Promise(resolve => setTimeout(resolve, 1000));
-      // return fetchPerson(personId);
-    }
-    throw error;
-  }
-};
+const fetchPerson = async (personId: number) =>
+  fetchTMDB<PersonResponse>(`https://api.themoviedb.org/3/person/${personId}?api_key=${apiKey}`);
 
 let processedPersons = 1;
-export const fetchPersonsData = async (personIds: number[], storeFunction = initPersonDB) => {
+export const fetchPersonsData = async (personIds: number[], storeFunction: (person: PersonData) => Promise<void>) => {
   for (const personId of personIds) {
     await new Promise(resolve => setTimeout(resolve, delay));
 
@@ -65,5 +46,3 @@ export const fetchPersonsData = async (personIds: number[], storeFunction = init
   }
   console.log("database filled with persons");
 };
-// const crewAndCastIds = await fetchMoviesData([2, 3]);
-// await fetchPersonsData(crewAndCastIds);
